DigitaleDeltaApi/function_app/__init__.py
import azure.functions as func
from DigitaleDeltaApi.function_app import app as flask_app

def main(req: func.HttpRequest, context: func.Context) -> func.HttpResponse:
    return func.WsgiMiddleware(flask_app.wsgi_app).handle(req, context)
