DigitaleDeltaApi/requirements.txt
flask
sqlalchemy
pyodbc
azure-functions
